requirements.txt
APScheduler==3.10.4
Flask==3.0.3
gunicorn==23.0.0
python-dotenv==1.0.1
requests==2.31.0
